Fix the bug of "QFont::setPixelSize:Pixel size <= 0 (-1)"   Modify the polish() function in ukui-styles/proxy-style/proxy-style.cpp file Fix the bug of fontsize types of int to double   Modify the Qt5UKUIPlatformTheme() function in qt5-ukui-platformtheme/qt5-ukui-platform-theme.cpp file Fix the bug of spinbox can't read the size of fontsize that int type   Modify the MainWindow() function in test/system-settings/mainwindow.cpp

+++ qt5-ukui-platformtheme/qt5-ukui-platform-theme.cpp
@@ -88,7 +88,7 @@ Qt5UKUIPlatformTheme::Qt5UKUIPlatformTheme(const QStringList &args)
                 }
             }
             if (key == "systemFontSize") {
-                int fontSize = settings->get("system-font-size").toString().toDouble();
+                double fontSize = settings->get("system-font-size").toString().toDouble();
                 QFontDatabase db;
                 if (fontSize > 0) {
                     QFont oldFont = QApplication::font();
